backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
pinecone-client
neo4j
httpx
pydantic
python-dotenv
motor
pymongo
google-generativeai
beautifulsoup4

